+++ src/Presentation/SmartStore.Web/Infrastructure/Installation/InstallDataSeeder.cs
@@ -21,6 +21,7 @@ using SmartStore.Core.Domain.Themes;
 using SmartStore.Core.Events;
 using SmartStore.Core.Infrastructure.DependencyManagement;
 using SmartStore.Core.Logging;
+using SmartStore.Core.Plugins;
 using SmartStore.Data;
 using SmartStore.Data.Setup;
 using SmartStore.Data.Utilities;
@@ -493,7 +494,7 @@ namespace SmartStore.Web.Infrastructure.Installation
 						null /* IWorkContext: not needed during install */,
 						rsResources,
 						languageService,
-						NullEventPublisher.Instance);
+						new PluginLocalizationDiffer(PluginFinder.Current));
 				}
 
 				return _locService;